Fix assertion handler signatures
With custom assertions enabled, the supplied long parameter leaves
SQLiteCpp's int overload undefined and causes link errors. Match the
public declaration in examples, tests, and documentation.

+++ examples/example2/src/main.cpp
@@ -21,7 +21,7 @@
 namespace SQLite
 {
 /// definition of the assertion handler enabled when SQLITECPP_ENABLE_ASSERT_HANDLER is defined in the project (CMakeList.txt)
-void assertion_failed(const char* apFile, const long apLine, const char* apFunc, const char* apExpr, const char* apMsg)
+void assertion_failed(const char* apFile, const int apLine, const char* apFunc, const char* apExpr, const char* apMsg)
 {
     // Print a message to the standard error output stream, and abort the program.
     std::cerr << apFile << ":" << apLine << ":" << " error: assertion failed (" << apExpr << ") in " << apFunc << "() with message \"" << apMsg << "\"\n";
